requirements.txt
Django==1.8.2
djangorestframework==3.4.4
MySQL-python==1.2.3
Pillow==3.2.0
